Move test utils to separate module.
